test(helpers): add centralized test helpers module for git-perf
- Introduced a new `test_helpers` module exposing utilities for test setup
- Provides hermetic git environment setup, git repo initialization, and run_git_command helpers
- Removed duplicate test setup code in various modules and replaced with usage of test_helpers
- Enabled `test_helpers` feature and integrated with benchmarks requiring test utilities
- Improves test code reuse, readability, and maintenance across unit tests and benchmarks
